rename: 관심 지역 패키지명 변경 및 이동
