+++ src/main/java/site/goldenticket/domain/user/wish/controller/WishRegionController.java
@@ -1,4 +1,4 @@
-package site.goldenticket.domain.wishRegion.controller;
+package site.goldenticket.domain.user.wish.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.goldenticket.common.response.CommonResponse;
-import site.goldenticket.domain.wishRegion.dto.WishRegionCreateRequest;
-import site.goldenticket.domain.wishRegion.dto.WishRegionListResponse;
-import site.goldenticket.domain.wishRegion.dto.WishRegionResponse;
-import site.goldenticket.domain.wishRegion.service.WishRegionService;
+import site.goldenticket.domain.user.wish.service.WishRegionService;
+import site.goldenticket.domain.user.wish.dto.WishRegionCreateRequest;
+import site.goldenticket.domain.user.wish.dto.WishRegionListResponse;
+import site.goldenticket.domain.user.wish.dto.WishRegionResponse;
 
 @RestController
 @RequiredArgsConstructor
